Few improvements to the versioning scheme. First changed to a global monotonically increased version number (this time really like subversion and its repo-level version number). Second returning info about what has been really deployed in remote deployment (by web service) because if you deploy foo you will end up with a package like foo-12.
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/ode/trunk@501522 13f79535-47bb-0310-9956-ffa450edef68

+++ bpel-dao/src/main/java/org/apache/ode/bpel/dao/ProcessDAO.java
@@ -18,11 +18,10 @@
  */
 package org.apache.ode.bpel.dao;
 
-import java.util.Collection;
+import org.apache.ode.bpel.common.CorrelationKey;
 
 import javax.xml.namespace.QName;
-
-import org.apache.ode.bpel.common.CorrelationKey;
+import java.util.Collection;
 
 /**
  * BPEL process data access objects. Contains references to active process instances ({@link ProcessInstanceDAO} and messages bound
@@ -48,7 +47,7 @@ public interface ProcessDAO {
      * 
      * @return version
      */
-    int getVersion();
+    long getVersion();
 
     /**
      * Get a message correlator instance.
